requirements.txt
numpy>=1.24.0
open3d>=0.18.0
matplotlib>=3.7.0
scipy>=1.10.0
scikit-learn>=1.3.0
Pillow>=10.0.0
plyfile>=0.7.4
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
